+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - 新增 `f-close-btn` 组件 `color` 属性
 - 废除 `f-button` 组件 `before-icon` `after-icon` 配置项
 - 新增 `f-button` 组件 `beforeIcon` `afterIcon` `loadingIcon` 插槽
+- 废除 `f-ink` 组件 `before-icon` `after-icon` 配置项
+- 新增 `f-ink` 组件 `beforeIcon` `afterIcon` 插槽
 
 ## 0.7.0-alpha.9 (2022-10-19)